requirements.txt
joblib==1.3.2
numpy==1.26.4
paho-mqtt==1.6.1
scikit-learn==1.3.1
scipy==1.12.0
Flask==3.0.2